requirements.txt
opencv-python
Pillow
tkinter
sv_ttk
moviepy
tqdm
numpy